+++ include/dxfeed_graal_cpp_api/event/market/TradeBase.hpp
@@ -193,7 +193,7 @@ class DXFCPP_EXPORT TradeBase : public MarketEvent, public LastingEvent {
      */
     void setSequence(std::int32_t sequence) noexcept {
         // TODO: Improve error handling [EN-8232]
-        assert(sequence >= 0 && sequence <= MAX_SEQUENCE);
+        assert(sequence >= 0 && static_cast<std::uint32_t>(sequence) <= MAX_SEQUENCE);
 
         tradeBaseData_.timeSequence = orOp(andOp(tradeBaseData_.timeSequence, ~MAX_SEQUENCE), sequence);
     }
